yaml file generator fix
The issue before was that the yaml file generator seemed to always produce the same yaml file made by taking the first option from each widget. This seemed to have something to do with the fact that the output of the widgets was being captured in a variable that was then used in another method. Removing the use of this middleman variable has the yaml generator working properly but tbh not really sure why.