Fixing tests (#79)
This is a mish-mash of fixes to get testing working. Includes a better (hopefully) way of loading config files. Ideally would come up with a more generic way for derivatives of POCS to do this.

Note that I've forced the tests with the mongo db only, unlike POCS, which does all three db types