Fixing pyro result (#88)
This is the fix that was implemented during a testing run. This isn't really a fix for the problem but at least lets us work. Seems like Pyro was trying to serialize an astropy Quantity but it was really unclear where that was coming from. Ultimately the `_async_wait` just wants a success boolean so here we capture the error and then set to True. Could have bad side effects.